+++ src/main/nextflow/co2footprint/CO2FootprintFactory.groovy
@@ -1,22 +1,18 @@
 package nextflow.co2footprint
 
-import nextflow.co2footprint.DataContainers.CIDataMatrix
-import nextflow.co2footprint.DataContainers.TDPDataMatrix
-import nextflow.co2footprint.Logging.LoggingAdapter
-
-import java.nio.file.Paths
-
+import groovy.transform.CompileStatic
 import groovy.transform.PackageScope
 import groovy.transform.PackageScopeTarget
-import groovy.transform.CompileStatic
-
 import groovy.util.logging.Slf4j
 
 import nextflow.Session
+import nextflow.co2footprint.DataContainers.CIDataMatrix
+import nextflow.co2footprint.DataContainers.TDPDataMatrix
+import nextflow.co2footprint.Logging.LoggingAdapter
 import nextflow.trace.TraceObserver
 import nextflow.trace.TraceObserverFactory
 
-
+import java.nio.file.Paths
 
 /**
  * Factory class for creating the CO2Footprint trace observer.