+++ src/main/nextflow/co2footprint/Records/CiRecordCollector.groovy
@@ -2,7 +2,7 @@ package nextflow.co2footprint.Records
 
 import groovy.util.logging.Slf4j
 import nextflow.co2footprint.CO2FootprintConfig
-import nextflow.co2footprint.Metrics.Converter
+import nextflow.co2footprint.Metrics.Duration
 import nextflow.exception.MissingValueException
 import nextflow.trace.TraceRecord
 
@@ -79,7 +79,7 @@ class CiRecordCollector {
      */
     void start(CO2FootprintConfig config=this.config, Integer delay=0, Integer period=1000*60*60) {
         if ( config.usesAPI() ) {
-            log.trace("Started periodically fetching the CI every ${Converter.toReadableTimeUnits(period, 'ms')}")
+            log.trace("Started periodically fetching the CI every ${new Duration(period, 'ms').toReadable()}")
             timer.scheduleAtFixedRate(new TimerTask() {
                 void run() {
                     add(config.get('ci').raw as CiRecord)