+++ src/test/nextflow/co2footprint/CO2FootprintConfigTest.groovy
@@ -1,15 +1,16 @@
 package nextflow.co2footprint
 
-import nextflow.co2footprint.DataContainers.DataMatrix
-import nextflow.co2footprint.DataContainers.TDPDataMatrix
-import nextflow.co2footprint.DataContainers.CIDataMatrix
-import spock.lang.Specification
-import groovy.util.logging.Slf4j
-import java.util.concurrent.ConcurrentHashMap
 import ch.qos.logback.classic.Logger
 import ch.qos.logback.classic.spi.ILoggingEvent
 import ch.qos.logback.core.read.ListAppender
+import groovy.util.logging.Slf4j
+import nextflow.co2footprint.DataContainers.CIDataMatrix
+import nextflow.co2footprint.DataContainers.DataMatrix
+import nextflow.co2footprint.DataContainers.TDPDataMatrix
 import org.slf4j.LoggerFactory
+import spock.lang.Specification
+
+import java.util.concurrent.ConcurrentHashMap
 
 @Slf4j
 class CO2FootprintConfigTest extends Specification {
@@ -136,7 +137,7 @@ class CO2FootprintConfigTest extends Specification {
         logger.addAppender(listAppender)
 
         when:
-        new CO2FootprintConfig(configMap, tdp, ci, processMap)
+        new CO2FootprintConfig(configMap, tdp, ci, processMap as Map)
 
         then:
         def logMessages = listAppender.list*.formattedMessage