+++ src/test/nextflow/co2footprint/CO2FootprintObserverTest.groovy
@@ -1,8 +1,10 @@
 package nextflow.co2footprint
 
+import groovy.util.logging.Slf4j
 import nextflow.NextflowMeta
 import nextflow.Session
 import nextflow.co2footprint.Records.CO2EquivalencesRecord
+import nextflow.co2footprint.TestHelpers.FileChecker
 import nextflow.executor.NopeExecutor
 import nextflow.processor.TaskHandler
 import nextflow.processor.TaskId
@@ -11,17 +13,13 @@ import nextflow.processor.TaskRun
 import nextflow.script.WorkflowMetadata
 import nextflow.trace.TraceObserver
 import nextflow.trace.TraceRecord
-
 import spock.lang.Shared
 import spock.lang.Specification
 
 import java.nio.file.Files
 import java.nio.file.Path
-
 import java.time.OffsetDateTime
 import java.util.concurrent.Executors
-import groovy.util.logging.Slf4j
-
 
 @Slf4j
 class CO2FootprintObserverTest extends Specification{
@@ -59,9 +57,9 @@ class CO2FootprintObserverTest extends Specification{
         return Mock(Session) {
             getConfig() >> [
                 co2footprint: [
-                    'traceFile': tracePath,
-                    'summaryFile': summaryPath,
-                    'reportFile': reportPath,
+                    'trace': ['enabled': true, 'file': tracePath],
+                    'summary': ['enabled': true, 'file': summaryPath],
+                    'report': ['enabled': true, 'file': reportPath],
                     'ci': ciValue
                 ]
             ]
@@ -113,15 +111,12 @@ class CO2FootprintObserverTest extends Specification{
         observer.onFlowCreate(session)
         observer.onProcessStart(handler, traceRecord)
         observer.onProcessComplete(handler, traceRecord)
+        observer.onFlowComplete()
 
         expect:
-        Double total_co2 = 0d
-        Double total_energy = 0d
-        observer.getCO2eRecords().values().each { co2Record ->
-            total_energy += co2Record.energy
-            total_co2 += co2Record.co2e
-        }
-        // With TDP = 11.41 (default global)
+        Double total_co2 =  observer.workflowStats.co2Record.store.co2e as Double
+        Double total_energy =  observer.workflowStats.co2Record.store.energy as Double
+        // With TDP = 11.45 (default global)
         // Energy consumption converted to Wh
         round(total_energy*1000) == 14.02
         // Total CO₂ in g (should reflect the CI value you set)
@@ -150,16 +145,11 @@ class CO2FootprintObserverTest extends Specification{
         observer.onFlowCreate(session)
         observer.onProcessStart(handler, traceRecord)
         observer.onProcessComplete(handler, traceRecord)
-
-        // Accumulate CO2
-        Double total_co2 = 0d
-        observer.getCO2eRecords().values().each { co2Record ->
-            total_co2 += co2Record.co2e
-        }
+        observer.onFlowComplete()
 
         CO2EquivalencesRecord co2EquivalencesRecord = observer
             .getCO2FootprintComputer()
-            .computeCO2footprintEquivalences(total_co2)
+            .computeCO2footprintEquivalences(observer.workflowStats.co2Record.store.co2e as Double)
 
         expect:
         // Values compared to result from www.green-algorithms.org (1h, 1core, TDP=11.45, CI:475)
@@ -183,12 +173,12 @@ class CO2FootprintObserverTest extends Specification{
         // Mock Session
         Session session = Mock(Session)
         session.getConfig() >> [
-                co2footprint:
-                        [
-                                'traceFile': tracePath,
-                                'summaryFile': summaryPath,
-                                'reportFile': reportPath
-                        ]
+            co2footprint:
+                [
+                    'trace': [enabled: true, file: tracePath],
+                    'summary': [enabled: true, file: summaryPath],
+                    'report': [enabled: true, file: reportPath]
+                ]
         ]
         session.getExecService() >> Executors.newFixedThreadPool(1)
         WorkflowMetadata meta = Mock(WorkflowMetadata)
@@ -209,6 +199,7 @@ class CO2FootprintObserverTest extends Specification{
         when:
         // Run necessary observer steps
         observer.onFlowCreate(session)
+        observer.onProcessStart(taskHandler, traceRecord)
         observer.onProcessComplete(taskHandler, traceRecord)
         observer.onFlowComplete()
 
@@ -230,10 +221,10 @@ class CO2FootprintObserverTest extends Specification{
         ]
         values == [
             '111', 'COMPLETED', '-', '14.02 Wh', '6.73 g', '-', '480 gCO₂e/kWh', '100 %',
-            '7 GB', '3600s', '1', '11.41 W', 'Unknown model', '11.41 Wh', '2.61 Wh',
+            '7 GB', '1h', '1', '11.41 W', 'Unknown model', '11.41 Wh', '2.61 Wh',
         ] // GA: CO₂e is 6.94g with CI of 475 gCO₂eq/kWh
 
-        fileChecker.compareChecksums(tracePath, '43c74a1981bb0a8c7694a97d697e22ae')
+        fileChecker.compareChecksums(tracePath, 'a9face6b58356811b2f77c87f6542b70')
 
 
         // Check Summary File
@@ -250,7 +241,7 @@ class CO2FootprintObserverTest extends Specification{
         fileChecker.runChecks(
             reportPath,
             [
-                1057: '    window.options = [' +
+                1005: '    window.options = [' +
                         '{"option":"ci","value":"480.0"},'+
                         '{"option":"ciMarket","value":null},' +
                         '{"option":"customCpuTdpFile","value":null},' +
@@ -263,7 +254,7 @@ class CO2FootprintObserverTest extends Specification{
                         "{\"option\":\"reportFile\",\"value\":\"${reportPath}\"}," +
                         "{\"option\":\"summaryFile\",\"value\":\"${summaryPath}\"}," +
                         "{\"option\":\"traceFile\",\"value\":\"${tracePath}\"}];",
-                1108: '          ' +
+                1056: '          ' +
                         "<span id=\"workflow_start\">${time.format('dd-MMM-YYYY HH:mm:ss')}</span>" +
                         " - <span id=\"workflow_complete\">${time.format('dd-MMM-YYYY HH:mm:ss')}</span>"
             ]